重构，将添加节点关系和存入数据库分离

+++ src/main/java/cn/edu/fudan/se/multidependency/model/node/Node.java
@@ -9,9 +9,9 @@ public interface Node extends Serializable {
 	
 	void setId(Long id);
 	
-	Integer getEntityId();
+	Long getEntityId();
 	
-	void setEntityId(Integer entityId);
+	void setEntityId(Long entityId);
 
 	Map<String, Object> getProperties();
