+++ src/main/java/cn/edu/fudan/se/multidependency/model/node/code/Type.java
@@ -21,7 +21,7 @@ public class Type implements Node {
 	
 	private String aliasName;
 
-    private Integer entityId;
+    private Long entityId;
 	private static final long serialVersionUID = 6805501035295416590L;
 	
 	public String getTypeName() {
@@ -42,10 +42,10 @@ public class Type implements Node {
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public Integer getEntityId() {
+	public Long getEntityId() {
 		return entityId;
 	}
-	public void setEntityId(Integer entityId) {
+	public void setEntityId(Long entityId) {
 		this.entityId = entityId;
 	}
