+++ src/main/java/cn/edu/fudan/se/multidependency/model/node/testcase/Scenario.java
@@ -22,7 +22,7 @@ public class Scenario implements Node {
     @GeneratedValue
     private Long id;
     
-    private Integer entityId;
+    private Long entityId;
     
 	@Override
 	public Long getId() {
@@ -35,12 +35,12 @@ public class Scenario implements Node {
 	}
 
 	@Override
-	public Integer getEntityId() {
+	public Long getEntityId() {
 		return entityId;
 	}
 
 	@Override
-	public void setEntityId(Integer entityId) {
+	public void setEntityId(Long entityId) {
 		this.entityId = entityId;
 	}
